+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## v1.1.1
 - Added reply for coins that are not found
 - Removed user-agent spoof since upstream API have changed and is now working without it
 - Removed BTC pairing for the price of Bitcoin (BTC) since it will always be 1.0